Extract excerpt query runner partial
